% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bandsift}
\alias{bandsift}
\title{Band-sift material editing}
\usage{
bandsift(
  im,
  freq,
  amp,
  sign,
  strength,
  log_epsilon = 1e-04,
  filter_epsilon = 0.01
)
}
\arguments{
\item{im}{an image}

\item{freq}{which scale to edit. either a numeric vector or a string ("A", "H", or "L")}

\item{amp}{which amplitude to edit. either "A", "H", or "L"}

\item{sign}{which sign to edit. either "A", "+", or "-"}

\item{strength}{a scaling factor}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}
}
\value{
an image
}
\description{
Band-sift material editing
}
\examples{
\donttest{
im = bandsift(regatta, freq = "H", amp = "H", sign = "+", strength = 2)
pplot(im)
}
}
