% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fft_lowpass_kernel}
\alias{fft_lowpass_kernel}
\title{Get lowpass kernel}
\usage{
fft_lowpass_kernel(kernel, height, width, cutoff, order = 3)
}
\arguments{
\item{kernel}{either "ideal", "gauss", "butterworth", or "dc"}
}
\description{
Get lowpass kernel
}
