% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fft_phase_scramble}
\alias{fft_phase_scramble}
\title{fft phase scrambling}
\usage{
fft_phase_scramble(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
fft phase scrambling
}
\examples{
pplot(fft_phase_scramble(regatta))
}
