% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_tricolor}
\alias{force_tricolor}
\title{Fprce an image to have three and only three color channels}
\usage{
force_tricolor(im)
}
\arguments{
\item{im}{an image}
}
\value{
an image
}
\description{
Fprce an image to have three and only three color channels
}
\examples{
R = get_R(regatta)
im = force_tricolor(R)
}
