% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gf_decompose}
\alias{gf_decompose}
\title{Scale-space decomposition by the guided filter}
\usage{
gf_decompose(im, log_epsilon = 1e-04, filter_epsilon = 0.01)
}
\arguments{
\item{im}{an image}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}
}
\value{
an image
}
\description{
Scale-space decomposition by the guided filter
}
\examples{
\donttest{
im = gf_decompose(regatta)
}
}
