% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gf_reconstruct}
\alias{gf_reconstruct}
\title{Reconstruct the original image from decomposed data}
\usage{
gf_reconstruct(dec)
}
\arguments{
\item{dec}{decomposed data}
}
\value{
an image
}
\description{
Reconstruct the original image from decomposed data
}
\examples{
\donttest{
im1 = regatta
dec = gf_decompose(im1)
im2 = gf_reconstruct(dec)
im_diff(im1, im2) # small difference means successful reconstruction
}
}
