% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_cx}
\alias{im_cx}
\title{X coordinate of image center}
\usage{
im_cx(im)
}
\arguments{
\item{im}{an image}
}
\value{
X coordinate of image center
}
\description{
X coordinate of image center
}
\examples{
im_cx(regatta)
}
