% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_cy}
\alias{im_cy}
\title{Y coordinate of image center}
\usage{
im_cy(im)
}
\arguments{
\item{im}{an image}
}
\value{
Y coordinate of image center
}
\description{
Y coordinate of image center
}
\examples{
im_cy(regatta)
}
