% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_diff}
\alias{im_diff}
\title{L2 distance of pixel value}
\usage{
im_diff(im1, im2)
}
\arguments{
\item{im1}{an image}

\item{im2}{an image}
}
\value{
mean squared error
}
\description{
L2 distance of pixel value
}
\examples{
im_diff(regatta, regatta^2)
}
