% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_load}
\alias{im_load}
\title{Load image from file or URL}
\usage{
im_load(file)
}
\arguments{
\item{file}{path to file or URL}
}
\value{
an array of image data
}
\description{
Load image from file or URL
}
\examples{
\dontrun{
# load an image from disk
im = im_load("path/to/your/image.jpg")
plot(im)
# load an image from URL
im = im_load("http://placehold.jp/150x150.png")
}
}
