% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{im_nchannel}
\alias{im_nchannel}
\title{The number of color channel of an image}
\usage{
im_nchannel(im)
}
\arguments{
\item{im}{an image}
}
\description{
The number of color channel of an image
}
\examples{
im_nchannel(regatta)
}
