% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_inside_ellipse}
\alias{is_inside_ellipse}
\title{Check if a point is inside or outside an ellipse}
\usage{
is_inside_ellipse(x, y, h, k, a, b)
}
\arguments{
\item{x}{x coordinate of the point}

\item{y}{y coordinate of the point}

\item{h}{x coordinate of the ellipse center}

\item{k}{y coordinate of the ellipse center}

\item{a}{semi-major axis x}

\item{b}{semi-major axis y}
}
\value{
mean squared error
}
\description{
Check if a point is inside or outside an ellipse
}
\examples{
is_inside_ellipse(x = 0, y = 0, h = 0, k = 0, a = 4, b = 2) #TRUE
is_inside_ellipse(x = 5, y = 0, h = 0, k = 0, a = 4, b = 2) #FALSE
}
