% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{medit}
\alias{medit}
\title{Band-sift material editing}
\usage{
medit(im, effect, scale, params, log_epsilon = 1e-04, filter_epsilon = 0.01)
}
\arguments{
\item{im}{an image}

\item{effect}{either "oily", "matte", "glossy", "stain", "shadow", "aging", "aging2", "smooth", or "smooth2"}

\item{scale}{a scaling factor}

\item{params}{a list. freq, amp, sign, scale}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}
}
\value{
an image
}
\description{
Band-sift material editing
}
\examples{
\donttest{
im = im_resize_scale(regatta, 0.6) # use a smaller image for speed

pplot(medit(im, effect = "oily"))
pplot(medit(im, effect = "oily", scale = 4))

oily = list(freq = "H", amp = "H", sign = "+", scale = 2)
pplot(medit(im, params = oily)) # oily effect
smooth = list(freq = "H", amp = "H", sign = "-", scale = 0.25)
pplot(medit(im, params = list(oily, smooth))) # oily and smooth effects
}
}
