% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nimg}
\alias{nimg}
\title{Create a nimg object}
\usage{
nimg(im)
}
\arguments{
\item{im}{a 3 dimensional numeric array}
}
\value{
an object of class nimg
}
\description{
nimg is a class for storing image data. Image data is stored as a 3 dimensional array with y, x, c format.
y and x are spatial coordinates, and c is a color dimension.
}
\examples{
nimg(array(1,c(128,256,3)))
}
