% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_channel}
\alias{rep_channel}
\title{Replicate an image along color dimension}
\usage{
rep_channel(im, n = 3)
}
\arguments{
\item{im}{an image}

\item{n}{number of iteration}
}
\description{
Replicate an image along color dimension
}
