% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_color}
\alias{split_color}
\title{Split color channel}
\usage{
split_color(im)
}
\arguments{
\item{im}{an image}
}
\value{
an list of images
}
\description{
Split color channel
}
\examples{
split_color(regatta) # a list of R, G, and B color channels
}
