% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{medit}
\alias{medit}
\title{Band-sift material editing}
\usage{
medit(im, effect, strength, params, log_epsilon = 1e-04, filter_epsilon = 0.01)
}
\arguments{
\item{im}{an image}

\item{effect}{either "oily", "matte", "glossy", "shadow", "shadow2", "aging", "aging2", "smooth", or "smooth2"}

\item{strength}{a scaling factor}

\item{params}{a list. freq, amp, sign, strength}

\item{log_epsilon}{offset for log transformation}

\item{filter_epsilon}{epsilon parameter}
}
\value{
an image
}
\description{
Band-sift material editing
}
\examples{
\donttest{
im = medit(regatta, "oily")
pplot(im)
im = medit(regatta, params = list(freq = "H", amp = "H", sign = "+", strength = 2))
}
}
